Streaming/requirements.txt
python-dateutil
tweepy
certifi
elasticsearch>=1.0.0,< 2.0.0